[test] Refactor opengl and vulkan cpp aot tests (#5812)
And move the macro control to cmake. There're a bit more to cleanup in
the cmake but this PR is getting too large, will leave those to followups.
