+++ tests/cpp/aot/vulkan/kernel_aot_test.cpp
@@ -1,0 +1,57 @@
+#include "gtest/gtest.h"
+#include "taichi/rhi/vulkan/vulkan_device.h"
+#include "taichi/rhi/vulkan/vulkan_device_creator.h"
+#include "taichi/rhi/vulkan/vulkan_loader.h"
+#include "tests/cpp/aot/gfx_utils.h"
+
+using namespace taichi;
+using namespace lang;
+
+TEST(GfxAotTest, VulkanDenseField) {
+  // Otherwise will segfault on macOS VM,
+  // where Vulkan is installed but no devices are present
+  if (!vulkan::is_vulkan_api_available()) {
+    return;
+  }
+
+  // Create Taichi Device for computation
+  lang::vulkan::VulkanDeviceCreator::Params evd_params;
+  evd_params.api_version = std::nullopt;
+  auto embedded_device =
+      std::make_unique<taichi::lang::vulkan::VulkanDeviceCreator>(evd_params);
+
+  aot_test_utils::run_dense_field_kernel(Arch::vulkan,
+                                         embedded_device->device());
+}
+
+TEST(GfxAotTest, VulkanKernelTest2) {
+  // Otherwise will segfault on macOS VM,
+  // where Vulkan is installed but no devices are present
+  if (!vulkan::is_vulkan_api_available()) {
+    return;
+  }
+
+  // Create Taichi Device for computation
+  lang::vulkan::VulkanDeviceCreator::Params evd_params;
+  evd_params.api_version = std::nullopt;
+  auto embedded_device =
+      std::make_unique<taichi::lang::vulkan::VulkanDeviceCreator>(evd_params);
+
+  aot_test_utils::run_kernel_test2(Arch::vulkan, embedded_device->device());
+}
+
+TEST(GfxAotTest, VulkanKernelTest1) {
+  // Otherwise will segfault on macOS VM,
+  // where Vulkan is installed but no devices are present
+  if (!vulkan::is_vulkan_api_available()) {
+    return;
+  }
+
+  // Create Taichi Device for computation
+  lang::vulkan::VulkanDeviceCreator::Params evd_params;
+  evd_params.api_version = std::nullopt;
+  auto embedded_device =
+      std::make_unique<taichi::lang::vulkan::VulkanDeviceCreator>(evd_params);
+
+  aot_test_utils::run_kernel_test1(Arch::vulkan, embedded_device->device());
+}
